+++ plugins/opencart4/upload/extension/payxcommerce/admin/language/en-gb/payment/payxcommerce.php
@@ -2,3 +2,4 @@
 $_['heading_title'] = 'PayXCommerce';
 $_['text_success'] = 'Success: PayXCommerce settings updated.';
 $_['error_permission'] = 'Warning: You do not have permission to modify PayXCommerce.';
+$_['help_public_brand'] = 'Shown to customers in checkout text. Use {brand} in title, description, and button text to reuse this value.';